refactor challenge two to fit challenge three[ch #164858466]
